+++ src/main/java/ru/trainee/creditmanager/dto/loanOffer/LoanOfferResponseShortDTO.java
@@ -4,6 +4,8 @@ import lombok.*;
 import ru.trainee.creditmanager.dto.bank.BankResponseShortDTO;
 import ru.trainee.creditmanager.dto.creditType.CreditTypeResponseShortDTO;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode
 @Builder
 @AllArgsConstructor
@@ -15,7 +17,7 @@ public class LoanOfferResponseShortDTO {
     Long id;
     BankResponseShortDTO bank;
     CreditTypeResponseShortDTO creditType;
-    Integer sumOfCredit;
+    BigDecimal sumOfCredit;
     boolean isActive;
     boolean isAccepted;
 }
